+++ examples/inverse/plot_custom_inverse_solver.py
@@ -73,10 +73,10 @@ def apply_solver(solver, evoked, forward, noise_cov, loose=0.2, depth=0.8):
         The forward solution.
     noise_cov : instance of Covariance
         The noise covariance.
-    loose : float in [0, 1] | 'auto' | None
+    loose : float in [0, 1] | 'auto'
         Value that weights the source variances of the dipole components
         that are parallel (tangential) to the cortical surface. If loose
-        is 0 or None then the solution is computed with fixed orientation.
+        is 0 then the solution is computed with fixed orientation.
         If loose is 1, it corresponds to free orientations.
         The default value ('auto') is set to 0.2 for surface-oriented source
         space and set to 1.0 for volumic or discrete source space.