+++ examples/inverse/plot_mixed_source_space_inverse.py
@@ -119,8 +119,7 @@ lambda2 = 1.0 / snr ** 2
 
 # Compute inverse operator
 inverse_operator = make_inverse_operator(evoked.info, fwd, noise_cov,
-                                         loose=None, depth=None,
-                                         fixed=False)
+                                         depth=None, fixed=False)
 
 stcs = apply_inverse(evoked, inverse_operator, lambda2, inv_method,
                      pick_ori=None)